+++ package.json
@@ -16,25 +16,17 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@chakra-ui/pro-theme": "^0.10.0",
-    "@chakra-ui/react": "2.8.2",
-    "@emotion/react": "^11.0.0",
-    "@emotion/styled": "^11.0.0",
     "@hookform/resolvers": "^3.3.2",
     "@paralleldrive/cuid2": "^2.2.2",
     "@radix-ui/react-alert-dialog": "^1.0.5",
-    "@radix-ui/react-avatar": "^1.0.4",
-    "@radix-ui/react-collapsible": "^1.0.3",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-hover-card": "^1.0.7",
     "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-radio-group": "^1.1.3",
-    "@radix-ui/react-scroll-area": "^1.0.5",
-    "@radix-ui/react-select": "^2.0.0",
-    "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-toggle": "^1.0.3",
+    "@radix-ui/react-toggle-group": "^1.0.4",
     "@tanstack/react-query": "^4.18.0",
     "@trpc/client": "10.43.6",
     "@trpc/react-query": "10.43.6",
@@ -46,13 +38,11 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "fast-jwt": "^3.3.1",
-    "framer-motion": ">=4.0.0",
     "ioredis": "^5.3.2",
     "lodash.mapvalues": "^4.6.0",
     "lucide-react": "^0.292.0",
     "morgan": "^1.10.0",
     "react": "^18.2.0",
-    "react-div-100vh": "^0.7.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.48.2",
     "react-router-dom": "^6.19.0",
