+++ src/client/components/icons.tsx
@@ -1,3 +1,5 @@
+import { cn } from '~/utils/misc';
+
 type IconProps = React.HTMLAttributes<SVGElement>;
 
 export const Icons = {
@@ -81,7 +83,7 @@ export const Icons = {
   ),
   diamond: (props: IconProps) => (
     <svg
-      fill="none"
+      fill="currentColor"
       viewBox="0 0 32 39"
       xmlns="http://www.w3.org/2000/svg"
       {...props}
@@ -89,7 +91,6 @@ export const Icons = {
       <path
         d="m15.106 1.6596c-0.9699 1.4388-3.0068 4.2991-3.4966 4.9099-0.1709 0.21327-0.9201 1.1606-1.6647 2.1051-1.5286 1.9388-3.243 3.9845-4.6424 5.5397-1.9425 2.1585-3.2926 3.6253-4.0974 4.4516l-0.8286 0.8509 0.48932 0.5396c0.26909 0.2968 1.271 1.3847 2.2266 2.4176 4.6787 5.0575 8.2622 9.5078 11.794 14.647 0.5968 0.8684 1.1059 1.5788 1.1313 1.5788 0.0256 0 0.3555-0.4612 0.7333-1.0248 3.5478-5.2928 8.3462-11.206 13.645-16.816 0.0306-0.0322 0.3248-0.3399 0.6537-0.6835l0.5983-0.6247-1.0425-1.0898c-5.1333-5.3666-10.283-11.675-13.641-16.71-0.447-0.6703-0.8495-1.2594-0.8944-1.3091-0.045-0.049746-0.4788 0.49868-0.9642 1.2187z"
         clipRule="evenodd"
-        fill="#A1312D"
         fillRule="evenodd"
       />
     </svg>
