--- src/client/components/suit-icon.tsx
@@ -1,119 +1,0 @@
-import { Icon } from '@chakra-ui/react';
-
-interface SuitIconProps {
-  opacity?: string;
-  isMini?: boolean;
-  color: string;
-  suit: 'heart' | 'diamond' | 'club' | 'spade';
-}
-
-const miniTransform = 'scale(0.9)';
-
-export function HeartIcon({
-  color,
-  opacity = '1',
-  isMini,
-}: Omit<SuitIconProps, 'suit'>) {
-  return (
-    <Icon
-      color={color}
-      viewBox="0 0 13 12"
-      width="13px"
-      height="12px"
-      transform={isMini ? miniTransform : ''}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M2.60217 0.0532025C1.93219 0.190011 1.40691 0.47507 0.897352 0.978354C0.479624 1.39092 0.251421 1.76952 0.096011 2.30777C-0.0323764 2.7524 -0.0319281 3.6404 0.0969076 4.11521C0.431734 5.34918 1.18148 6.37718 2.97873 8.06651C4.56797 9.56032 5.18789 10.2202 5.91387 11.1907C6.36735 11.7969 6.45457 11.9181 6.47165 11.9653C6.49357 12.0261 6.47931 12.0415 6.70417 11.7144C6.93466 11.379 7.45587 10.7029 7.82204 10.2641C8.15328 9.86726 8.95343 9.06868 10.1042 7.98646C11.8095 6.38286 12.573 5.33073 12.8871 4.15191C13.0296 3.61672 13.0384 2.79405 12.9067 2.31361C12.5166 0.891096 11.0911 -0.0854191 9.53774 0.00590757C9.31357 0.0191211 9.02574 0.0577547 8.89813 0.0918764C8.3472 0.238958 7.77146 0.579771 7.4134 0.970699C7.22811 1.17297 6.75691 1.78672 6.60916 2.0182C6.56938 2.08052 6.52121 2.12584 6.50209 2.11891C6.48298 2.11198 6.33511 1.92506 6.17354 1.70349C5.4538 0.716459 4.94828 0.326135 4.07618 0.0842624C3.76817 -0.00114236 2.95245 -0.0183442 2.60217 0.0532025Z"
-        fill="currentColor"
-        fillOpacity={opacity}
-      />
-    </Icon>
-  );
-}
-
-export function DiamondIcon({
-  color,
-  opacity = '1',
-  isMini,
-}: Omit<SuitIconProps, 'suit'>) {
-  return (
-    <Icon
-      color={color}
-      viewBox="0 0 12 14"
-      width="12px"
-      height="14px"
-      transform={isMini ? miniTransform : ''}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M5.65227 0.447078C5.28007 0.973521 4.49845 2.02012 4.31051 2.24362C4.24493 2.32165 3.95744 2.66826 3.6717 3.01386C3.08514 3.72326 2.42727 4.47178 1.89027 5.04082C1.14488 5.83061 0.626795 6.36731 0.317957 6.66965L0 6.98099L0.187766 7.17842C0.291024 7.28702 0.675505 7.68508 1.04217 8.06303C2.83754 9.91356 4.21262 11.5419 5.56802 13.4223C5.79704 13.7401 5.99239 14 6.00215 14C6.01195 14 6.13856 13.8312 6.28354 13.625C7.64495 11.6884 9.48621 9.52463 11.5196 7.47192C11.5313 7.46013 11.6442 7.34755 11.7704 7.22184L12 6.99324L11.6 6.59449C9.63017 4.63085 7.65399 2.32251 6.36545 0.480159C6.19392 0.234896 6.03948 0.0193489 6.02227 0.00114679C6.00501 -0.0170553 5.83851 0.183614 5.65227 0.447078Z"
-        fill="currentColor"
-        fillOpacity={opacity}
-      />
-    </Icon>
-  );
-}
-
-export function ClubIcon({
-  color,
-  opacity = '1',
-  isMini,
-}: Omit<SuitIconProps, 'suit'>) {
-  return (
-    <Icon
-      color={color}
-      viewBox="0 0 14 14"
-      width="14px"
-      height="14px"
-      transform={isMini ? miniTransform : ''}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M6.5372 0.0246256C6.43948 0.0345846 6.2826 0.0608773 6.18857 0.0830565C5.9997 0.127607 5.58359 0.260562 5.5445 0.288827C5.5308 0.298738 5.49362 0.314543 5.46187 0.323948C5.39148 0.344829 5.07767 0.522623 4.9302 0.625172C4.56594 0.878524 4.21988 1.22719 3.96773 1.5949C3.92418 1.6584 3.88854 1.71477 3.88854 1.72013C3.88854 1.72549 3.85719 1.78178 3.81884 1.84522C3.45065 2.45452 3.28328 3.36656 3.40366 4.10776C3.44841 4.38334 3.58131 4.80241 3.69916 5.03965L3.78281 5.20804L3.48083 5.22353C3.31473 5.23207 3.0892 5.25591 2.97963 5.27652C2.77552 5.31494 2.37572 5.43199 2.22014 5.49887C2.06027 5.56759 1.95879 5.61684 1.94622 5.63177C1.93937 5.63993 1.90575 5.65492 1.87151 5.66509C1.83727 5.67527 1.80366 5.69025 1.79681 5.69841C1.77587 5.72331 1.62556 5.82145 1.60835 5.82145C1.57411 5.82145 1.25273 6.08067 1.0726 6.25361C0.889669 6.42921 0.55173 6.83486 0.55173 6.87886C0.55173 6.89077 0.537711 6.91482 0.520603 6.93231C0.460939 6.99322 0.253733 7.42271 0.179651 7.63897C-0.0415748 8.28483 -0.0602508 9.12545 0.133434 9.71845C0.205573 9.93932 0.24173 10.0324 0.3189 10.1957C0.364769 10.2928 0.402321 10.3821 0.402321 10.3941C0.402321 10.4062 0.413526 10.4161 0.427222 10.4161C0.440918 10.4161 0.452124 10.4298 0.452124 10.4467C0.452124 10.4635 0.491344 10.5348 0.539279 10.6051C0.587215 10.6754 0.626435 10.7429 0.626435 10.7549C0.626435 10.767 0.636146 10.7769 0.648024 10.7769C0.659903 10.7769 0.70114 10.8229 0.739662 10.8791C0.840638 11.0266 1.18196 11.3783 1.22405 11.3783C1.23777 11.3783 1.24897 11.3875 1.24897 11.3988C1.24897 11.4278 1.72225 11.7632 1.76312 11.7632C1.78164 11.7632 1.79688 11.7713 1.79701 11.7812C1.79711 11.7911 1.88954 11.8413 2.00245 11.8927C2.11532 11.9442 2.22701 11.9955 2.25064 12.0068C2.30161 12.0313 2.54057 12.1098 2.68081 12.1482C2.94114 12.2195 3.31944 12.2669 3.63442 12.2676C3.94813 12.2683 4.41411 12.2173 4.52353 12.1703C4.55092 12.1585 4.61816 12.1388 4.67294 12.1265C4.98339 12.0569 5.50378 11.81 5.81914 11.5827C6.02761 11.4325 6.30691 11.1727 6.46585 10.9811C6.52064 10.9151 6.57106 10.8557 6.57791 10.8491C6.60588 10.822 6.70242 10.6816 6.70242 10.6679C6.70242 10.6598 6.71923 10.6369 6.73977 10.6171C6.77296 10.585 6.77705 10.5872 6.77653 10.6369C6.7762 10.6676 6.76432 10.736 6.75011 10.7889C6.73591 10.8418 6.70892 10.9501 6.69012 11.0295C6.64746 11.2095 6.59721 11.3895 6.52881 11.6068C6.49967 11.6994 6.46615 11.8077 6.45435 11.8474C6.4221 11.9557 6.37658 12.0793 6.33915 12.1601C6.32075 12.1998 6.29849 12.2621 6.28965 12.2986C6.28081 12.3351 6.2636 12.371 6.25143 12.3782C6.23925 12.3855 6.22929 12.4076 6.22929 12.4274C6.22929 12.4472 6.20625 12.5142 6.17812 12.5764C6.14998 12.6385 6.11746 12.711 6.10585 12.7374C6.03065 12.9087 5.65471 13.6337 5.56529 13.7798C5.50592 13.8768 5.45734 13.9663 5.45734 13.9786C5.45734 13.9922 6.05814 14.0006 6.98777 14C8.45738 13.999 8.61566 13.9919 8.52965 13.9315C8.51762 13.923 8.48694 13.8782 8.46144 13.8319C8.43594 13.7856 8.35835 13.6449 8.28897 13.5192C8.12786 13.2273 8.01996 13.012 7.89705 12.7374C7.8146 12.5532 7.77413 12.4577 7.73392 12.3525C7.70864 12.2864 7.67358 12.2052 7.65602 12.1721C7.63844 12.139 7.62403 12.0934 7.62393 12.0707C7.62385 12.048 7.60796 11.9939 7.58867 11.9504C7.48856 11.7252 7.22535 10.7683 7.22535 10.6296C7.22535 10.584 7.24298 10.5985 7.31664 10.705C7.36687 10.7776 7.41448 10.8424 7.42247 10.8491C7.43047 10.8557 7.48127 10.9152 7.53538 10.9814C7.65167 11.1236 7.94092 11.4013 8.06874 11.4935C8.11847 11.5293 8.17348 11.5718 8.19099 11.5877C8.23364 11.6266 8.63506 11.8612 8.7444 11.9112C9.12875 12.0867 9.40479 12.171 9.7933 12.2312C10.2371 12.3 10.8885 12.2679 11.2843 12.1577C11.4742 12.1049 11.7005 12.0316 11.7518 12.0062C11.7755 11.9946 11.89 11.9412 12.0065 11.8876C12.2351 11.7824 12.5227 11.599 12.7298 11.4264C12.8611 11.317 13.1043 11.0804 13.1675 11.0006C13.3896 10.7203 13.4859 10.5753 13.6184 10.3214C13.7042 10.1569 13.7745 10.0048 13.7745 9.98344C13.7745 9.96206 13.7846 9.9343 13.797 9.92172C13.8248 9.89345 13.8888 9.65939 13.9512 9.35761C14.0133 9.05699 14.0168 8.44906 13.9579 8.15484C13.9133 7.93143 13.7857 7.50466 13.7351 7.40912C13.7175 7.37604 13.6773 7.29485 13.6457 7.2287C13.4914 6.90604 13.2481 6.57119 12.9613 6.28661C12.7762 6.10302 12.5218 5.89362 12.4839 5.89362C12.4678 5.89362 12.4547 5.88402 12.4547 5.87231C12.4547 5.84883 12.1247 5.65306 12.0851 5.65306C12.0715 5.65306 12.0567 5.64496 12.0521 5.63502C12.0385 5.60546 11.7804 5.49134 11.5582 5.41669C11.1919 5.2936 10.7797 5.22002 10.4563 5.22004C10.3228 5.22006 10.2135 5.21566 10.2135 5.21025C10.2135 5.20486 10.2407 5.14804 10.2739 5.08396C10.4485 4.74653 10.5498 4.42683 10.6116 4.01729C10.6614 3.68732 10.6211 2.98882 10.5374 2.73032C10.5245 2.69063 10.4909 2.58247 10.4628 2.49C10.3993 2.28144 10.3804 2.23675 10.2281 1.93694C10.1078 1.70028 10.0202 1.55944 9.9832 1.54359C9.97292 1.53916 9.96453 1.52446 9.96453 1.5109C9.96453 1.4733 9.72437 1.18997 9.5484 1.01997C9.36796 0.845664 9.01737 0.575762 8.86261 0.492024C8.09989 0.0792798 7.39403 -0.0625997 6.5372 0.0246256Z"
-        fill="currentColor"
-        fillOpacity={opacity}
-      />
-    </Icon>
-  );
-}
-
-export function SpadeIcon({
-  color,
-  opacity = '1',
-  isMini,
-}: Omit<SuitIconProps, 'suit'>) {
-  return (
-    <Icon
-      color={color}
-      viewBox="0 0 13 14"
-      width="13px"
-      height="14px"
-      transform={isMini ? miniTransform : ''}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M6.19871 0.410651C5.58052 1.29889 4.96681 2.00735 4.04673 2.89491C1.7833 5.07832 1.31747 5.57434 0.81294 6.33841C0.562301 6.71803 0.331905 7.18766 0.191749 7.6047C-0.10275 8.48087 -0.0549879 9.50888 0.314141 10.238C0.948502 11.491 2.41079 12.2136 3.82077 11.9709C4.71622 11.8168 5.39122 11.3708 6.03254 10.5096C6.18497 10.3049 6.31817 10.1374 6.32854 10.1374C6.40935 10.1374 5.92624 11.8316 5.65231 12.5088C5.54923 12.7637 5.13331 13.5939 5.00604 13.7988C4.95795 13.8763 4.91852 13.9532 4.91844 13.9698C4.91836 13.9864 5.63166 14 6.50357 14C7.37548 14 8.08886 13.9911 8.08886 13.9803C8.08886 13.9695 7.98655 13.7777 7.86152 13.5542C7.3465 12.6334 6.86103 11.2695 6.68832 10.2581L6.6643 10.1173L6.75669 10.2348C7.51873 11.2043 7.79274 11.4557 8.38109 11.7248C8.84578 11.9373 9.16857 12.0036 9.73513 12.003C10.4759 12.0022 10.9908 11.848 11.612 11.4406C12.5648 10.8159 13.0542 9.82835 12.9952 8.64962C12.9578 7.9014 12.7419 7.23055 12.3161 6.53939C11.7878 5.68157 11.2511 5.0856 9.619 3.54427C8.95439 2.91655 8.22036 2.18322 7.98781 1.91457C7.55279 1.41195 6.84417 0.500014 6.63958 0.179458C6.57597 0.0797551 6.51475 -0.000957074 6.50357 8.57541e-06C6.49239 0.00101446 6.3552 0.185816 6.19871 0.410651Z"
-        fill="currentColor"
-        fillOpacity={opacity}
-      />
-    </Icon>
-  );
-}
-
-export function SuitIcon({
-  color,
-  suit,
-  opacity = '1',
-  isMini,
-}: SuitIconProps) {
-  const props = { color, opacity, isMini };
-  if (suit === 'heart') return <HeartIcon {...props} />;
-  if (suit === 'diamond') return <DiamondIcon {...props} />;
-  if (suit === 'club') return <ClubIcon {...props} />;
-  return <SpadeIcon {...props} />;
-}
