+++ src/client/components/vote-picker.tsx
@@ -1,10 +1,10 @@
-import { SimpleGrid, Text, useBreakpointValue, VStack } from '@chakra-ui/react';
+import * as ToggleGroupPrimitive from '@radix-ui/react-toggle-group';
 
-import { BigCard } from '~/components/big-card';
 import { useGame } from '~/hooks/use-game';
 import { usePlayer } from '~/hooks/use-player';
 import { api } from '~/utils/api';
-import { getSuitFromCardIndex } from '~/utils/misc';
+
+import { Icons } from './icons';
 
 export function VotePicker() {
   const game = useGame();
@@ -12,35 +12,25 @@ export function VotePicker() {
 
   const vote = api.player.vote.useMutation();
 
-  const columns = useBreakpointValue({
-    base: 4,
-    lg: game.votingSystem.values.length,
-  });
-
   return (
-    <VStack spacing={4} alignItems="flex-start">
-      <Text
-        as="h2"
-        fontSize="lg"
-        fontWeight="bold"
-        opacity={player.isSpectator ? 0.4 : 1}
-      >
-        Pick your points 👇
-      </Text>
-      <SimpleGrid columns={columns} spacing={6}>
-        {game.votingSystem.values.map((value, i) => (
-          <BigCard
-            key={i}
-            points={value}
-            suit={getSuitFromCardIndex(i)}
-            isSelected={player.vote === value}
-            onClick={async () =>
-              await vote.mutateAsync(player.vote === value ? undefined : value)
-            }
-            disabled={game.isRevealed || player.isSpectator}
-          />
-        ))}
-      </SimpleGrid>
-    </VStack>
+    <ToggleGroupPrimitive.Root
+      type="single"
+      className="flex flex-wrap gap-4"
+      aria-label="Choose vote"
+      value={player.vote}
+      onValueChange={async (value) => await vote.mutateAsync(value)}
+    >
+      {game.votingSystem.values.map((value) => (
+        <ToggleGroupPrimitive.Item
+          key={value}
+          value={value}
+          className="relative flex h-20 w-14 items-center justify-center rounded-lg border-4 border-primary text-primary transition-colors hover:opacity-90 disabled:pointer-events-none disabled:opacity-50 data-[state=on]:bg-primary data-[state=disabled]:text-primary-foreground data-[state=on]:text-primary-foreground"
+        >
+          <Icons.diamond className="absolute left-1 top-1 h-3 opacity-40" />
+          <p className="text-2xl font-bold ">{value}</p>
+          <Icons.diamond className="absolute bottom-1 right-1 h-3 opacity-40" />
+        </ToggleGroupPrimitive.Item>
+      ))}
+    </ToggleGroupPrimitive.Root>
   );
 }
